backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
cssselect2==0.8.0
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
frozenlist==1.6.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.4.3
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pyzmq==26.4.0
razorpay==1.4.2
requests==2.32.3
setuptools==78.1.1
six==1.17.0
stack-data==0.6.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.2
traitlets==5.14.3
twilio==9.5.2
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
yarl==1.20.0
zopfli==0.2.3.post1
gunicorn